Remove unneccessary differences between the original WLED pixel art converter and art.html.
This was done in an attempt to fix a bug where solid black (#000000) becomes a noisy black (#010000, #000100, etc.).
It turns out this is a bug in Firefox in Ubuntu 23.10, and only happens when the art.html is loaded from the device.
The black is not noisy if art.html is loaded from the local filesystem, loaded from the device or locally in Chrome, nor
does it happen if loaded from the device in Firefox in Windows 11.